Make player.html video more responsive
Until some proper technology is used, this change would make the video a bit more responsive.

The change is small enough, so that browser CPU utilization is kept below 7% on average (while the normal is about 3-4% on i5-4430). 50ms delay would make the video more stable, but CPU usage would be double.